some refactoring: packages added, method initCategoriesFill() moved to MainActivity from ExpensesFragment
