+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
             android:value="ru.loftschool.loftblogmoneytracker.database.model.Expenses,
                             ru.loftschool.loftblogmoneytracker.database.model.Categories" />
 
-        <activity android:name=".SplashActivity">
+        <activity android:name=".ui.activities.SplashActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
@@ -24,7 +24,7 @@
         </activity>
 
         <activity
-            android:name=".MainActivity_"
+            android:name=".ui.activities.MainActivity_"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="ru.loftschool.loftblogmoneytracker.MAINACTIVITY" />
@@ -33,10 +33,10 @@
         </activity>
 
         <activity
-            android:name=".AddExpenseActivity_">
+            android:name=".ui.activities.AddExpenseActivity_">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".MainActivity"/>
+                android:value=".ui.activities.MainActivity"/>
         </activity>
     </application>
