+++ app/src/main/java/ru/loftschool/loftblogmoneytracker/adapters/CategoriesAdapter.java
@@ -1,4 +1,4 @@
-package ru.loftschool.loftblogmoneytracker;
+package ru.loftschool.loftblogmoneytracker.adapters;
 
 import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
@@ -8,6 +8,7 @@ import android.widget.TextView;
 
 import java.util.List;
 
+import ru.loftschool.loftblogmoneytracker.R;
 import ru.loftschool.loftblogmoneytracker.database.model.Categories;
 
 public class CategoriesAdapter extends RecyclerView.Adapter<CategoriesAdapter.CardViewCategoryHolder> {
