+++ app/src/main/java/ru/loftschool/loftblogmoneytracker/ui/fragments/ExpensesFragment.java
@@ -1,4 +1,4 @@
-package ru.loftschool.loftblogmoneytracker;
+package ru.loftschool.loftblogmoneytracker.ui.fragments;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -21,8 +21,11 @@ import org.androidannotations.annotations.res.StringRes;
 
 import java.util.List;
 
+import ru.loftschool.loftblogmoneytracker.adapters.ExpensesAdapter;
+import ru.loftschool.loftblogmoneytracker.R;
 import ru.loftschool.loftblogmoneytracker.database.model.Categories;
 import ru.loftschool.loftblogmoneytracker.database.model.Expenses;
+import ru.loftschool.loftblogmoneytracker.ui.activities.AddExpenseActivity_;
 
 @EFragment(R.layout.expenses_fragment)
 public class ExpensesFragment extends Fragment {
@@ -50,8 +53,6 @@ public class ExpensesFragment extends Fragment {
         recyclerView.setLayoutManager(linearLayoutManager);
 //      to avoid an error "recyclerview No adapter attached; skipping layout" set a blank adapter for the recyclerView
         recyclerView.setAdapter(new ExpensesAdapter());
-
-        initialCategoriesFill();
 
         Bundle args = getArguments();
         if (args != null){
@@ -95,14 +96,5 @@ public class ExpensesFragment extends Fragment {
 
     private List<Expenses> getDataList(){
         return new Select().from(Expenses.class).execute();
-    }
-
-    private void initialCategoriesFill() {
-        if (new Select().from(Categories.class).execute().size() == 0) {
-            new Categories("Social").save();
-            new Categories("Fun").save();
-            new Categories("Clothes").save();
-            new Categories("Food").save();
-        }
     }
 }
